Populate GameOverviewFragment with stubbed data from the Game Repository.
Let's first create an adapter so we can manage the RecyclerView. Glide is used to make sure the image is loaded as efficient as possible.
Once the Adapter is in place, we can create a ViewModel where we hold the games in a LiveData object which we can observe in the Fragment. The games are being populated statically until we implemented Retrofit to make the actual API call(s).
The last step is glue them together in the OverviewFragment.